+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  * Make Symfony 2.7 LTS as a minimum requirement.
  * Remove deprecated `IsoCodesConstraintValidator` interface.
+ * Remove deprecated `ZipCode` country option values.
 
 * 1.1.0 (2015-??-??)